requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
get-annotations==0.1.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
protobuf==5.29.3
psycopg2==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-mock==3.14.0
python-dotenv==1.0.1
python-telegram-bot==21.10
requests==2.32.3
rsa==4.9
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
Telethon==1.38.1
typing_extensions==4.12.2
urllib3==2.3.0
yandex-cloud-ml-sdk==0.2.4
yandexcloud==0.330.0
